+++ src/hooks/useValidateParams.ts
@@ -1,0 +1,6 @@
+import { useSearchParams } from 'react-router-dom';
+
+export const useValidateParams = (allowed: string[]) => {
+  const [params] = useSearchParams();
+  return !Array.from(params.keys()).some((key) => !allowed.includes(key));
+};
