+++ src/main.tsx
@@ -5,9 +5,10 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import { store } from './store/store';
 
-import App from './App';
 import AboutPage from './pages/AboutPage/AboutPage';
 import NotFoundPage from './pages/NotFoundPage/NotFoundPage';
+import HomePage from './pages/HomePage/HomePage';
+
 import './index.css';
 import GlowCursor from './components/Cursor/Cursor';
 
@@ -23,7 +24,7 @@ createRoot(rootEl).render(
       <ThemeProvider>
         <BrowserRouter basename="/rss-school">
           <Routes>
-            <Route path="/" element={<App />} />
+            <Route path="/" element={<HomePage />} />
             <Route path="/about" element={<AboutPage />} />
             <Route path="*" element={<NotFoundPage />} />
           </Routes>
